+++ web/modules/custom/hoeringsportal_citizen_proposal_archiving/src/Plugin/AdvancedQueue/JobType/ArchiveCitizenProposalJob.php
@@ -2,10 +2,10 @@
 
 namespace Drupal\hoeringsportal_citizen_proposal_archiving\Plugin\AdvancedQueue\JobType;
 
+use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
 use Drupal\advancedqueue\Job;
 use Drupal\advancedqueue\JobResult;
 use Drupal\advancedqueue\Plugin\AdvancedQueue\JobType\JobTypeBase;
-use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
 use Drupal\hoeringsportal_citizen_proposal_archiving\Helper\Helper;
 use Symfony\Component\DependencyInjection\ContainerInterface;