+++ web/modules/custom/hoeringsportal_forms/src/Plugin/views/filter/NodeIndexNid.php
@@ -83,7 +83,7 @@ final class NodeIndexNid extends ManyToOne {
   /**
    * {@inheritdoc}
    */
-  public function init(ViewExecutable $view, DisplayPluginBase $display, array &$options = NULL) {
+  public function init(ViewExecutable $view, DisplayPluginBase $display, ?array &$options = NULL) {
     parent::init($view, $display, $options);
 
     if (!empty($this->definition['node'])) {